backend/python-ml/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
numpy==1.26.3
pandas==2.1.4
scikit-learn==1.4.0
opencv-python-headless==4.9.0.80
tensorflow==2.15.0
pillow==10.2.0
python-multipart==0.0.6
python-dotenv==1.0.0
joblib==1.3.2
